+++ switchbot_client/devices/__init__.py
@@ -1,0 +1,2 @@
+from .physical import *
+from .remote import *
